Initial commit WIP
Trying to figure out how the architecture of this works. Not tried using
an ORM with TypeScript before and it would be great if typing worked
nicely with it. TypeOrm looks like it should work nicely with it.

Also trying to use TypeScript on the front with React Native, which is
not a new thing for me but I'm interested in trying to share the non
display stuff (ie redux sagas and selectors) between web and app.